Simplify API internals
