--- sweetest/sweetest-api/src/main/java/com/mysugr/sweetest/framework/accessor/StatesAccessor.kt
@@ -1,3 +1,0 @@
-package com.mysugr.sweetest.framework.accessor
-
-class StatesAccessor(accessor: BaseAccessor)
