--- sweetest/sweetest-api/src/main/java/com/mysugr/sweetest/framework/accessor/StepsAccessor.kt
@@ -1,5 +1,0 @@
-package com.mysugr.sweetest.framework.accessor
-
-import com.mysugr.sweetest.framework.context.TestContext
-
-class StepsAccessor(testContext: TestContext) : BaseAccessor(testContext)
