--- sweetest/sweetest-api/src/main/java/com/mysugr/sweetest/framework/accessor/TestAccessor.kt
@@ -1,5 +1,0 @@
-package com.mysugr.sweetest.framework.accessor
-
-import com.mysugr.sweetest.framework.context.TestContext
-
-class TestAccessor(testContext: TestContext) : BaseAccessor(testContext)
