+++ sweetest/sweetest-api/src/main/java/com/mysugr/sweetest/framework/base/BaseJUnitTest.kt
@@ -2,29 +2,30 @@ package com.mysugr.sweetest.framework.base
 
 import com.mysugr.sweetest.framework.build.TestBuilder
 import com.mysugr.sweetest.framework.configuration.ModuleTestingConfiguration
+import com.mysugr.sweetest.framework.context.TestContext
 import org.junit.After
 import org.junit.Before
 
 abstract class BaseJUnitTest @Deprecated(
     "No module configuration needed anymore.",
     ReplaceWith("BaseJUnitTest()", imports = ["BaseJUnitTest"])
-) constructor(private val moduleTestingConfiguration: ModuleTestingConfiguration? = null) : TestingAccessor {
+) constructor(private val moduleTestingConfiguration: ModuleTestingConfiguration? = null) : CommonBase(TestContext()) {
 
     constructor() : this(moduleTestingConfiguration = null)
 
-    open fun configure() = TestBuilder(moduleTestingConfiguration)
+    open fun configure() = TestBuilder(testContext, moduleTestingConfiguration)
 
-    override val accessor = configure().build()
-    protected val dependencies = accessor.dependencies
-    protected val delegates = accessor.delegates
+    init {
+        configure().build()
+    }
 
     @Before
     fun junitBefore() {
-        accessor.testContext.workflowController.run()
+        testContext.workflowController.run()
     }
 
     @After
     fun junitAfter() {
-        accessor.testContext.workflowController.finish()
+        testContext.workflowController.finish()
     }
 }
